% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gg}
\alias{gg}
\title{Large igraph for example usage}
\format{An object of class \code{"list"} with an igraph object under \code{gg$g}}
\usage{
data(gg)
}
\description{
Large igraph for example usage
}
\keyword{datasets}
