% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust2igraph.R
\name{hclust2igraph}
\alias{hclust2igraph}
\title{Convert an hclust to an igraph}
\usage{
hclust2igraph(hc)
}
\arguments{
\item{hc}{a hclust object.}
}
\value{
An igraph object.
}
\description{
Function for converting a hclust object to an igraph.
}
\examples{
hc <- hclust(dist(USArrests), "ave")
gg <- hclust2igraph(hc)                                 

}
\seealso{
\code{\link[stats:hclust]{hclust}}
}
