% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normggSize.R
\name{normggSize}
\alias{normggSize}
\title{Normalization of the igraph node sizes}
\usage{
normggSize(gg, mint, maxt)
}
\arguments{
\item{gg}{An igraph object with a nodeSize component set <igraph>.}

\item{mint}{The target bottom limit for the size normalization <numeric>.}

\item{maxt}{The target upper limit for the size normalization <numeric>.}
}
\value{
an igraph object with its nodeSize and legNodeSize components normalized
}
\description{
Returns the given igraph object with the node sizes normalized to the
range indicated by the user. If you also want to normalize the legend scales,
it is necessary to use this function after aplying \code{\link[TreeAndLeaf]{formatTree}} or 
\code{\link[RedeR]{att.setv}}
}
\examples{
data(gg)
gg <- gg$g
gg <- normggSize (gg, 150, 50)

}
\seealso{
\code{\link[TreeAndLeaf]{treeAndLeaf}}

\code{\link[RedeR]{addGraph}}

\code{\link[igraph]{V}}
}
