% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo2igraph.R
\name{phylo2igraph}
\alias{phylo2igraph}
\title{Convert a phylo object to an igraph}
\usage{
phylo2igraph(phy)
}
\arguments{
\item{phy}{A phylo object with labeled tips <phylo>.}
}
\value{
An igraph object.
}
\description{
Function for converting a phylo object to an igraph.
}
\examples{
phy <- ape::rtree(10, tip.label = c(1:10))
gg <- phylo2igraph(phy)

}
