% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_species-data.R
\docType{data}
\name{phylo_species}
\alias{phylo_species}
\title{Species metadata from STRING-db v11}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(phylo_species)
}
\description{
Species metadata used in a phylo object.
}
\references{
Szklarczyk D. et al. STRING v11: protein-protein association networks with 
increased coverage, supporting functional discovery in genome-wide 
experimental datasets. Nucleic Acids Res. 47:D607-613, 2019.
}
\keyword{datasets}
