% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_tree-data.R
\docType{data}
\name{phylo_tree}
\alias{phylo_tree}
\title{Species tree from STRING-db v11}
\format{
An object of class \code{"phylo"}
}
\usage{
data(phylo_tree)
}
\description{
A phylo object for the species tree available from the STRING-db v11.
}
\references{
Szklarczyk D. et al. STRING v11: protein-protein association networks with 
increased coverage, supporting functional discovery in genome-wide 
experimental datasets. Nucleic Acids Res. 47:D607-613, 2019.
}
\keyword{phylotree}
