% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTree.R
\name{formatTree}
\alias{formatTree}
\title{Formatting of an igraph to be shown in RedeR}
\usage{
formatTree(gg, dataframe, ..., theme = "SSGreens", cleanalias = FALSE)
}
\arguments{
\item{gg}{An igraph object <igraph>.}

\item{dataframe}{A dataframe containing at least rownames set to be the same
used to create the igraph, a column with the data to be shown as node sizes
and a column with the data to create the color gradient <dataframe>.}

\item{theme}{Selects the predefined theme to be the base of the igraph formatting.
The first character represents the size of the tree - S for small, M for medium and L for large.
See \code{\link{treeAndLeaf}} for more information about tree sizes.
The second character represents the type of color palette - S for sequential, D for divergent.
The remaining characters represent the color palette - Greens and Reds for sequential palettes
and PuOr for a Purple/Orange divergent palette. Options are: SSGreens, SSReds, SDPuOr,
MSGreens, MSReds, MDPuOr, LSGreens, LSReds, LDPuOr (default = "SSGreens") <string>.}

\item{cleanalias}{A logical that removes the node aliases when set to TRUE (default = FALSE) <logical>.}

\item{nodeSize}{The name of the column for size differentiation (must be a column of given "dataframe") <string>.}

\item{nszleg}{Number of size legend components <numeric>.}

\item{nodeColor}{The name of the column color differentiation (must be a column of given "dataframe") <string>.}

\item{palqt}{How many color quantiles the data will be divided into <numeric>.}

\item{colpal}{Which RColorBrewer palette will be used <string>.}

\item{nodeFontSize}{The alias font size <numeric>.}

\item{nodeFontColor}{The alias font color <string>.}

\item{edgeWidth}{The edge width <numeric>.}

\item{edgeColor}{The edge color <numeric>.}
}
\value{
An igraph object with standard formatting for RedeR application.
}
\description{
Applies formatting patterns to an igraph object according to predefined themes
and/or user choices. This formatting is used for plotting on the RedeR app interface.
You may choose among using the original themes, personalizing just some of the parameters
or setting all of them.
}
\examples{
hc <- hclust(dist(USArrests), "ave")
gg <- hclust2igraph(hc)
gg <- formatTree(gg = gg,
                 dataframe = USArrests,
                 theme = "SSReds",
                 nodeSize = "UrbanPop",
                 nodeColor = "Murder")

}
\seealso{
\code{\link[RedeR]{addGraph}}

\code{\link{treeAndLeaf}}
}
