% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTree.R
\name{formatTree}
\alias{formatTree}
\title{A theme function for formatting of an igraph to be shown in RedeR}
\usage{
formatTree(gg, theme = 1, cleanalias = FALSE)
}
\arguments{
\item{gg}{An igraph object generated by either \code{\link{hclust2igraph}}
or \code{\link{phylo2igraph}}<igraph>.}

\item{theme}{An integer ranging from 1 to 5 with desired theme. \cr
Options are: \cr
1- A clean black and blue theme, great for higher levels of user 
customization. \cr
2- A theme with a palette of greens. \cr
3- A theme with a palette of blues. \cr
4- A theme with a palette of purples. \cr
5- A theme with a palette of reds. \cr
For custom formatting, see \code{\link[RedeR]{addGraph}} for accepted 
parameters <integer>. \cr}

\item{cleanalias}{A logical that removes the node aliases when set to 
TRUE (default = FALSE) <logical>.}
}
\value{
An igraph object with standard formatting for RedeR application.
}
\description{
Applies formatting patterns to an igraph object 
according to predefined themes.
This formatting is used for plotting on the RedeR app interface.
}
\examples{
hc <- hclust(dist(USArrests), "ave")
gg <- hclust2igraph(hc)
gg <- formatTree(gg = gg,
                 theme = 5)
         
}
\seealso{
\code{\link[RedeR]{addGraph}}

\code{\link{treeAndLeaf}}
}
