% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_species-data.R
\docType{data}
\name{phylo_species}
\alias{phylo_species}
\title{Species metadata for the strings tree from STRING-db v11.0}
\format{An object of class \code{"data.frame"}}
\usage{
data(phylo_species)
}
\description{
Species metadata for phylo_tree object from STRING-db website.
Available visiting download page under Accessory Data.
Version: 11.0
}
\references{
STRING: functional protein association networks.
Retrieved August 15, 2019, from
https://string-db.org/
}
\keyword{datasets}
