% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_tree-data.R
\docType{data}
\name{phylo_tree}
\alias{phylo_tree}
\title{Species tree from STRING-db v11.0}
\format{An object of class \code{"phylo"}}
\usage{
data(phylo_tree)
}
\description{
Species tree from STRING-db website.
Available visiting download page under Accessory Data.
Version: 11.0
}
\references{
STRING: functional protein association networks.
Retrieved August 15, 2019, from
https://string-db.org/
}
\keyword{phylotree}
