% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdata-data.R
\docType{data}
\name{spdata}
\alias{spdata}
\title{Genome statistical data from eukaryotes with complete genome}
\format{An object of class \code{"data.frame"}}
\usage{
data(spdata)
}
\description{
Data from NCBI Genome database showing statistical results for
eukaryotes with complete genome sequenced.
The list of organisms with complete genome is available on KEGG 
Organisms website and NCBI data was scraped using RCrawler package.
Taxonomy ids were obtained using NCBI Tax Identifier.
}
\examples{
data(spdata)
View(spdata)
}
\references{
Genome - NCBI.
Retrieved August 15, 2019 from
https://www.ncbi.nlm.nih.gov/genome

Taxonomy - NCBI. Tax Identifier.
Retrieved August 15, 2019 from
https://www.ncbi.nlm.nih.gov/taxonomy

KEGG Organisms: Complete Genomes.
Retrieved August 15, 2019 from
https://www.genome.jp/kegg/catalog/org_list.html
}
\keyword{datasets}
