% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeAndLeaf.R
\name{treeAndLeaf}
\alias{treeAndLeaf}
\title{Initial layout creation for the TreeAndLeaf.}
\usage{
treeAndLeaf(obj, gg, size = "small", showgraph = TRUE)
}
\arguments{
\item{obj}{An object of RedPort class, from RedeR package <RedPort>.}

\item{gg}{An hclust or igraph object, containing the dendrogram to
be reorganized <igraph><hclust>.}

\item{size}{The size of the dendrogram. "small" = less
than 100 nodes; "medium" = less than 500 nodes; "large" = 500+ nodes.
This is just a reference point, use the size that gives the best
results (default = "small") <string>.}

\item{showgraph}{Toggles the addition of the resulting layout to the
RedeR app and the relaxation process (default = TRUE) <logical>.}
}
\value{
The given igraph with nodes coordinates added.
}
\description{
Returns the given igraph with nodes coordinates added after setting of positions
and relaxation by the force based algorithm implemented in the RedeR package.
}
\examples{
\dontrun{
rdp <- RedeR::RedPort()
RedeR::calld(rdp)
hc <- hclust(dist(USArrests), "ave")
treeAndLeaf(rdp, hc)
}
}
\seealso{
\code{\link{formatTree}}

\code{\link[RedeR:addGraph]{addGraph}}

\code{\link[RedeR:relax]{relax}}
}
