% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeAndLeaf.R
\name{treeAndLeaf}
\alias{treeAndLeaf}
\title{Layout creation and plotting of the TreeAndLeaf in RedeR.}
\usage{
treeAndLeaf(obj, gg)
}
\arguments{
\item{obj}{An object of RedPort class, from RedeR package <RedPort>.}

\item{gg}{An igraph object generated by either \code{\link{hclust2igraph}}
or \code{\link{phylo2igraph}}<igraph>.}
}
\value{
Plotting of igraph in RedeR app and the given igraph with nodes
coordinates added.
}
\description{
Creates tree-and-leaf layouts and plots. It also returns the given igraph
with nodes coordinates added after setting of positions and relaxation by
the force based algorithm implemented in the RedeR package.
}
\examples{
library(RedeR)
rdp <- RedPort()
hc <- hclust(dist(USArrests), "ave")
gg <- hclust2igraph(hc)

\dontrun{
calld(rdp)
treeAndLeaf(rdp, gg)
}

}
\seealso{
\code{\link{formatTree}}

\code{\link[RedeR:addGraph]{addGraph}}

\code{\link[RedeR:relax]{relax}}
}
